feat: update to latest `mc.py` with `sampler.py` support - feat: add `mc` related cfg to `config.py` - fix: fix standalone `sampler` execution by moving to module dir - feat: add `sampler` module init - feat: shorten `run_simulation` api to `run_sim` - chore: remove legacy `monte_carlo.py` module - feat: add plotter for `sampler` param coverage visualization - chore: update api call for `run_sim` in `checker.py` - feat: add `sampler` support for `mc.py` - chore: update to latest api call in `main.py`
