+++ src/strw_amuse/core/__init__.py
@@ -2,10 +2,11 @@
 strw_amuse.core package for stellar dynamics simulations project using AMUSE.
 """
 
-from . import collision, helpers, stopping
+from . import collision, helpers, sampler, stopping
 
 __all__ = [
     'collision',
     'helpers',
+    'sampler',
     'stopping',
 ]
